:alien: Oh PHP... you and your encoding issues
Data appears to be actually CP-1252 (Windows 1252) encoded in the DB,
but the column has a character set of latin1_swedish_ci. When that's
loaded and the MySQL driver respects this information, you get back
garbage.

So, we have a custom field to intercept the data loading. Turn the
"utf-8 string" back into bytes by encoding it with the (presumably)
correct windows-1252 encoding, and then turn it into an UTF-8 string
again.